cf_delete_instance_tp/requirements.txt
functions-framework==3.*
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2